pyproject.toml
[tool.poetry]
name = "casemk"
version = "0.1.0"
description = "Divided storage case generator for 3D printing"
readme = "README.md"
packages = [{include = "casemk"}]

[tool.poetry.dependencies]
python = "^3.8"
solidpython = ">=1.1.0"
setuptools = "*"  # required by solidpython
pyyaml = {version = ">=6.0", optional = true}

[tool.poetry.group.dev.dependencies]
# Add dev deps here if needed

[tool.poetry.extras]
yaml = ["pyyaml"]

[tool.poetry.scripts]
casemk = "casemk.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
